test(validator): move diagnostics tests to matching test file
Move TestEmitDiagnosticBlock, TestTailLines, and TestCollectNCCLWorkerDiagnostics
into nccl_all_reduce_bw_constraint_test.go, next to the functions they exercise,
instead of nccl_preflight_tcpxo_test.go. Keeps test discovery aligned with the
source layout. Addresses maintainer review nit.
