+++ validators/performance/nccl_preflight_tcpxo_test.go
@@ -15,17 +15,9 @@
 package main
 
 import (
-	"bytes"
-	"context"
-	"log/slog"
-	"strings"
 	"testing"
 
 	"github.com/NVIDIA/aicr/pkg/recipe"
-	corev1 "k8s.io/api/core/v1"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/runtime"
-	"k8s.io/client-go/kubernetes/fake"
 )
 
 func TestGKETCPXOPreflightApplies(t *testing.T) {
@@ -61,147 +53,6 @@ func TestGKETCPXOPreflightApplies(t *testing.T) {
 		t.Run(tt.name, func(t *testing.T) {
 			if got := gkeTCPXOPreflightApplies(tt.variant, tt.accelerator, tt.service); got != tt.want {
 				t.Errorf("gkeTCPXOPreflightApplies() = %v, want %v", got, tt.want)
-			}
-		})
-	}
-}
-
-func TestEmitDiagnosticBlock(t *testing.T) {
-	tests := []struct {
-		name         string
-		label        string
-		block        string
-		wantLines    int // number of "diagnostics" records emitted
-		wantContains []string
-	}{
-		{
-			name:         "multi-line block emits one record per line",
-			label:        "worker diagnostics",
-			block:        "line one\nline two\nline three",
-			wantLines:    3,
-			wantContains: []string{"line one", "line two", "line three", "worker diagnostics"},
-		},
-		{
-			name:         "empty block emits a single (empty) marker",
-			label:        "launcher logs",
-			block:        "   \n  ",
-			wantLines:    1,
-			wantContains: []string{"(empty)", "launcher logs"},
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			var buf bytes.Buffer
-			prev := slog.Default()
-			slog.SetDefault(slog.New(slog.NewTextHandler(&buf, &slog.HandlerOptions{Level: slog.LevelError})))
-			defer slog.SetDefault(prev)
-
-			emitDiagnosticBlock(tt.label, tt.block)
-
-			out := buf.String()
-			if got := strings.Count(out, "msg=diagnostics"); got != tt.wantLines {
-				t.Errorf("emitted %d diagnostic records, want %d\noutput:\n%s", got, tt.wantLines, out)
-			}
-			for _, want := range tt.wantContains {
-				if !strings.Contains(out, want) {
-					t.Errorf("output missing %q\noutput:\n%s", want, out)
-				}
-			}
-		})
-	}
-}
-
-func TestTailLines(t *testing.T) {
-	tests := []struct {
-		name string
-		in   string
-		n    int
-		want string
-	}{
-		{"fewer than n", "a\nb", 5, "a\nb"},
-		{"exactly n", "a\nb\nc", 3, "a\nb\nc"},
-		{"more than n keeps tail", "a\nb\nc\nd", 2, "c\nd"},
-		{"single line", "only", 3, "only"},
-		{"empty", "", 3, ""},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			if got := tailLines(tt.in, tt.n); got != tt.want {
-				t.Errorf("tailLines(%q, %d) = %q, want %q", tt.in, tt.n, got, tt.want)
-			}
-		})
-	}
-}
-
-func TestCollectNCCLWorkerDiagnostics(t *testing.T) {
-	const ns = "aicr-test"
-
-	workerLabels := map[string]string{
-		"jobset.sigs.k8s.io/jobset-name":        ncclTrainJobName,
-		"jobset.sigs.k8s.io/replicatedjob-name": nodeJobName,
-	}
-
-	failedWorker := &corev1.Pod{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      "nccl-all-reduce-tj-node-0-0-abcde",
-			Namespace: ns,
-			Labels:    workerLabels,
-		},
-		Status: corev1.PodStatus{
-			Phase: corev1.PodFailed,
-			// tcpxo-daemon is a native sidecar (init container).
-			InitContainerStatuses: []corev1.ContainerStatus{{
-				Name: "tcpxo-daemon",
-				State: corev1.ContainerState{
-					Terminated: &corev1.ContainerStateTerminated{
-						Reason:   "Error",
-						ExitCode: 137,
-						Message:  "fastrak init failed",
-					},
-				},
-			}},
-			ContainerStatuses: []corev1.ContainerStatus{{
-				Name: nodeJobName,
-				State: corev1.ContainerState{
-					Waiting: &corev1.ContainerStateWaiting{
-						Reason:  "CrashLoopBackOff",
-						Message: "back-off restarting",
-					},
-				},
-			}},
-		},
-	}
-
-	tests := []struct {
-		name         string
-		pods         []runtime.Object
-		wantContains []string
-	}{
-		{
-			name:         "no worker pods",
-			pods:         nil,
-			wantContains: []string{"no NCCL worker pods found"},
-		},
-		{
-			name: "worker with failed and waiting containers",
-			pods: []runtime.Object{failedWorker},
-			wantContains: []string{
-				failedWorker.Name,
-				"phase=Failed",
-				"container tcpxo-daemon: terminated reason=Error exitCode=137",
-				"fastrak init failed",
-				"container node: waiting reason=CrashLoopBackOff",
-			},
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			client := fake.NewClientset(tt.pods...)
-			got := collectNCCLWorkerDiagnostics(context.Background(), client, ns)
-			for _, want := range tt.wantContains {
-				if !strings.Contains(got, want) {
-					t.Errorf("diagnostics missing %q\nfull output:\n%s", want, got)
-				}
 			}
 		})
 	}
